Delete events, sync-unsync calendars

+++ app/controllers/api/v1/google_calendars_controller.rb
@@ -1,14 +1,22 @@
 class Api::V1::GoogleCalendarsController < ApiController
-  # before_action :google_auth, only: [:sync]
+  before_action :google_auth, only: [:sync]
+  before_action :check_for_params, only: [:unsync_calendar, :sync_calendar]
+  before_action :set_calendar, only: [:unsync_calendar, :sync_calendar]
 
   def sync
-    google_auth
     items = []
+    google_events_ids = []
     @accounts.each do |service|
       account = account(service.authorization.access_token)
       parser = GoogleCalendars.new(current_user, service, account)
       parser.import_calendars
       items << parser.items
+      parser.items.each { |item| google_events_ids << item.id }
+      local_events = Event.where('google_event_id is not NULL and events.user_id = ?', current_user.id)
+        .includes(:calendar)
+        .where(calendars: {sync_with_google: true})
+      result = local_events.pluck(:google_event_id) - google_events_ids
+      compare_events(result)
     end
     render json: {items: items}
   end
@@ -17,18 +25,43 @@ class Api::V1::GoogleCalendarsController < ApiController
     google_access_token = GoogleAccessToken.find_by_account account
     authorize google_access_token
     calendars_list = @service.list_calendar_lists.items.map {|calendar| calendar.id}
-    calendars = Calendar.where('title in (?) AND account = ?', calendars_list, account)
-    ActiveRecord::Base.transaction do
-      calendars.each do |calendar|
-        calendar.events.each do |event|
-          event.destroy
-        end
-        calendar.destroy
-      end
-    end
+    calendars = Calendar.where('google_calendar_id in (?) AND account = ?', calendars_list, account)
+    calendars.destroy_all
+  end
+
+  def unsync_calendar
+    @calendar.unsync! if @calendar
+    render nothing: true
+  end
+
+  def sync_calendar
+    @calendar.sync! if @calendar
+    render nothing: true
   end
 
   private
+
+  def compare_events(result)
+    result.each do |e|
+      event = Event.find_by_google_event_id(e)
+      event.destroy if event
+    end
+  end
+
+  def check_for_params
+    errors = []
+    %w(calendar account).each do |a|
+      errors << "Set #{a}" if params[a.to_sym].blank?
+    end
+    unless errors.empty?
+      render json: {errors: errors}, status: 403
+      return
+    end
+  end
+
+  def set_calendar
+    @calendar = Calendar.find_by(title: params[:calendar], account: params[:account])
+  end
 
   def account(access_token)
     uri = account_info_uri + access_token
@@ -44,11 +77,13 @@ class Api::V1::GoogleCalendarsController < ApiController
     end
   end
 
-  def authorize(google_access_token)
-    @google_oauth ||= Api::V1::GoogleOauthController.new
-    @google_oauth.refresh_token google_access_token if google_access_token.expired?
-    @client = Signet::OAuth2::Client.new(access_token: google_access_token.token)
-    @service ||= Google::Apis::CalendarV3::CalendarService.new
-    @service.authorization = @client
-  end
+  # def authorize(google_access_token)
+  #   puts "AUTHORIZATION"
+  #   @google_oauth ||= Api::V1::GoogleOauthController.new
+  #   puts "google oauth #{@google_oauth}"
+  #   @google_oauth.refresh_token google_access_token if google_access_token.expired?
+  #   @client = Signet::OAuth2::Client.new(access_token: google_access_token.token)
+  #   @service ||= Google::Apis::CalendarV3::CalendarService.new
+  #   @service.authorization = @client
+  # end
 end
